+++ src/main/java/faang/school/projectservice/controller/SubProjectController.java
@@ -1,9 +1,9 @@
 package faang.school.projectservice.controller;
 
-import faang.school.projectservice.dto.CreateSubProjectDto;
-import faang.school.projectservice.dto.SubProjectFilterDto;
-import faang.school.projectservice.dto.ProjectDto;
-import faang.school.projectservice.dto.UpdateSubProjectDto;
+import faang.school.projectservice.dto.subproject.CreateSubProjectDto;
+import faang.school.projectservice.dto.subproject.SubProjectFilterDto;
+import faang.school.projectservice.dto.project.ProjectDto;
+import faang.school.projectservice.dto.subproject.UpdateSubProjectDto;
 import faang.school.projectservice.service.ProjectService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -32,13 +33,13 @@ public class SubProjectController {
         return projectService.createSubProject(parentProjectId, createDto);
     }
 
-    @PatchMapping("/{projectId}")
+    @PutMapping("/{projectId}")
     public ProjectDto updateSubProject(@Positive @PathVariable Long projectId,
                                        @RequestBody UpdateSubProjectDto updateDto) {
         return projectService.updateSubProject(projectId, updateDto);
     }
 
-    @GetMapping("/{parentProjectId}/subprojects")
+    @PostMapping("/{parentProjectId}/subprojects")
     public List<ProjectDto> getFilteredSubProjects(@Positive @PathVariable Long parentProjectId,
                                                    @RequestBody SubProjectFilterDto filterDto) {
         return projectService.getFilteredSubProjects(parentProjectId, filterDto);
