+++ src/main/java/faang/school/projectservice/dto/project/ProjectDto.java
@@ -1,19 +1,18 @@
-package faang.school.projectservice.dto;
+package faang.school.projectservice.dto.project;
 
+import faang.school.projectservice.model.ProjectStatus;
 import faang.school.projectservice.model.ProjectVisibility;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Builder;
 
-import java.util.List;
-
 @Builder
-public record CreateSubProjectDto(
+public record ProjectDto(
         Long id,
         @NotBlank String name,
         @NotNull @Positive Long ownerId,
         @NotNull ProjectVisibility visibility,
-        List<StageDto> stages
+        @NotNull ProjectStatus status
 ) {
 }
