+++ src/main/java/faang/school/projectservice/dto/stage/StageDto.java
@@ -1,6 +1,5 @@
-package faang.school.projectservice.dto;
+package faang.school.projectservice.dto.stage;
 
-import faang.school.projectservice.model.stage.Stage;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
