+++ src/main/java/faang/school/projectservice/dto/subproject/CreateSubProjectDto.java
@@ -1,18 +1,20 @@
-package faang.school.projectservice.dto;
+package faang.school.projectservice.dto.subproject;
 
-import faang.school.projectservice.model.ProjectStatus;
+import faang.school.projectservice.dto.stage.StageDto;
 import faang.school.projectservice.model.ProjectVisibility;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Builder;
 
+import java.util.List;
+
 @Builder
-public record ProjectDto(
+public record CreateSubProjectDto(
         Long id,
         @NotBlank String name,
         @NotNull @Positive Long ownerId,
         @NotNull ProjectVisibility visibility,
-        @NotNull ProjectStatus status
+        List<StageDto> stages
 ) {
 }
