+++ src/main/java/faang/school/projectservice/filter/subproject/SubProjectFilter.java
@@ -1,6 +1,6 @@
-package faang.school.projectservice.filter;
+package faang.school.projectservice.filter.subproject;
 
-import faang.school.projectservice.dto.SubProjectFilterDto;
+import faang.school.projectservice.dto.subproject.SubProjectFilterDto;
 import faang.school.projectservice.model.Project;
 
 import java.util.List;
