+++ src/main/java/faang/school/projectservice/filter/subproject/SubProjectNameFilter.java
@@ -1,24 +1,24 @@
-package faang.school.projectservice.filter;
+package faang.school.projectservice.filter.subproject;
 
-import faang.school.projectservice.dto.SubProjectFilterDto;
+import faang.school.projectservice.dto.subproject.SubProjectFilterDto;
 import faang.school.projectservice.model.Project;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
-public class SubProjectStatusFilter implements SubProjectFilter {
+public class SubProjectNameFilter implements SubProjectFilter {
 
     @Override
     public boolean isApplicable(SubProjectFilterDto subProjectFilterDto) {
-        return subProjectFilterDto.status() != null;
+        return subProjectFilterDto.name() != null;
     }
 
     @Override
     public List<Project> apply(List<Project> projects,
                                SubProjectFilterDto subProjectFilterDto) {
         return projects.stream()
-                .filter(project -> project.getStatus().equals(subProjectFilterDto.status()))
+                .filter(project -> project.getName().equals(subProjectFilterDto.name()))
                 .toList();
     }
 }
