+++ src/main/java/faang/school/projectservice/mapper/StageMapper.java
@@ -1,6 +1,6 @@
 package faang.school.projectservice.mapper;
 
-import faang.school.projectservice.dto.StageDto;
+import faang.school.projectservice.dto.stage.StageDto;
 import faang.school.projectservice.model.stage.Stage;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
